Deepen RibbonLine: merge geometry, raycast, and per-frame uniform service into RibbonNode (#53)
Collapses ribbonLine.js, the node-ribbon parts of lineFactory.js, and
lineMaterialResolutionService.js into a single src/ribbonNode.ts module.
The Z layering constant lives in one private place (read by geometry,
raycast, and getPoint), and raycast reads a module-scope halfWidth
snapshot instead of poking material.uniforms — severing the cross-module
coupling that previously made hit-testing depend on the per-frame
uniform service having run.

Look loses the four register calls, the vestigial zOffset field, and
the manual unregister loop in dispose(). Material registration is now
a side effect of RibbonNode.create(), not a Look discipline.

Net: -530 lines, 3 files deleted, 1 new TS module. 229/229 tests pass.
Adds 5 new boundary tests: Z source-of-truth (getPoint(u).z and vertex
positions both equal NODE_Z_OFFSET) and registration lifecycle (create
adds, dispose removes, dispose idempotent, clearRibbonRegistry bulk).

Closes #53.
